+++ lib/babel-to-estree/attachComments.js
