+++ lib/babel-to-estree/convertComments.js
