+++ lib/babel-to-estree/convertTemplateType.js
