+++ lib/babel-to-estree/index.js
