+++ lib/babel-to-estree/toToken.js
