+++ package.json
@@ -10,19 +10,11 @@
     "type": "git",
     "url": "https://github.com/babel/babel-eslint.git"
   },
-  "dependencies": {
-    "@babel/code-frame": "7.0.0-beta.44",
-    "@babel/traverse": "7.0.0-beta.44",
-    "@babel/types": "7.0.0-beta.44",
-    "babylon": "7.0.0-beta.44",
-    "eslint-scope": "~3.7.1",
-    "eslint-visitor-keys": "^1.0.0"
-  },
   "scripts": {
     "test": "npm run lint && npm run test-only",
     "test-only": "mocha && mocha --require test/fixtures/preprocess-to-patch.js && mocha --require test/fixtures/use-eslint-old.js",
-    "lint": "eslint index.js babylon-to-espree test",
-    "fix": "eslint index.js babylon-to-espree test --fix",
+    "lint": "eslint index.js babel-to-estree test",
+    "fix": "npm run lint -- --fix",
     "precommit": "lint-staged",
     "preversion": "npm test",
     "changelog": "git log `git describe --tags --abbrev=0`..HEAD --pretty=format:' * %s (%an)' | grep -v 'Merge pull request'"
@@ -37,14 +29,16 @@
   },
   "homepage": "https://github.com/babel/babel-eslint",
   "peerDependencies": {
-    "@babel/code-frame": ">=7.0.0-beta.40",
-    "@babel/core": ">=7.0.0-beta.40",
-    "babylon": ">=7.0.0-beta.40",
+    "@babel/core": ">=7.0.0-beta.51",
     "eslint": ">=4.14.0"
   },
+  "dependencies": {
+    "@babel/core": ">=7.0.0-beta.51",
+    "eslint-scope": "~3.7.1",
+    "eslint-visitor-keys": "^1.0.0"
+  },
   "devDependencies": {
-    "@babel/core": "7.0.0-beta.40",
-    "babel-eslint": "^8.0.0",
+    "babel-eslint": "^8.2.3",
     "dedent": "^0.7.0",
     "eslint": "npm:eslint@4.13.1",
     "eslint-config-babel": "^7.0.1",
